+++ core/types/market.go
@@ -330,6 +330,10 @@ type InstrumentSpot struct {
 	Spot *Spot
 }
 
+func (InstrumentSpot) Type() ProductType {
+	return ProductTypeSpot
+}
+
 func (i InstrumentSpot) String() string {
 	return fmt.Sprintf(
 		"spot(%s)",
@@ -367,6 +371,10 @@ func (s Spot) String() string {
 
 type InstrumentFuture struct {
 	Future *Future
+}
+
+func (InstrumentFuture) Type() ProductType {
+	return ProductTypeFuture
 }
 
 func (i InstrumentFuture) String() string {
@@ -485,6 +493,26 @@ func (m *Market) GetAssets() ([]string, error) {
 	return m.TradableInstrument.Instrument.Product.getAssets()
 }
 
+func (m *Market) ProductType() ProductType {
+	return m.TradableInstrument.Instrument.Product.Type()
+}
+
+func (m *Market) GetFuture() *InstrumentFuture {
+	if m.ProductType() == ProductTypeFuture {
+		f, _ := m.TradableInstrument.Instrument.Product.(*InstrumentFuture)
+		return f
+	}
+	return nil
+}
+
+func (m *Market) GetSpot() *InstrumentSpot {
+	if m.ProductType() == ProductTypeSpot {
+		s, _ := m.TradableInstrument.Instrument.Product.(*InstrumentSpot)
+		return s
+	}
+	return nil
+}
+
 func (i InstrumentFuture) iIntoProto() interface{} {
 	return i.IntoProto()
 }
@@ -493,6 +521,7 @@ type iProto interface {
 	iIntoProto() interface{}
 	getAssets() ([]string, error)
 	String() string
+	Type() ProductType
 }
 
 type Instrument struct {
@@ -502,6 +531,7 @@ type Instrument struct {
 	Metadata *InstrumentMetadata
 	// Types that are valid to be assigned to Product:
 	//	*InstrumentFuture
+	//	*InstrumentSpot
 	Product iProto
 }